Polish workspace styles
