+++ src/styles.css
@@ -37,49 +37,107 @@ button:disabled {
   min-height: 100vh;
 }
 
-.app-shell {
+.workspace {
   max-width: 1200px;
   margin: 0 auto;
-  padding: 32px 24px 48px;
+  padding: 36px 24px 56px;
   display: flex;
   flex-direction: column;
-  gap: 32px;
+  gap: 40px;
 }
 
-header {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-header h1 {
-  margin: 0;
-  font-size: 2.75rem;
-  letter-spacing: -0.025em;
-}
-
-header p {
-  margin: 0;
-  max-width: 720px;
-  color: rgba(226, 232, 240, 0.8);
-}
-
-main {
-  display: grid;
-  gap: 24px;
-}
-
-.column {
+.workspace__header {
   display: flex;
   flex-direction: column;
   gap: 24px;
 }
 
 @media (min-width: 1024px) {
-  main {
-    grid-template-columns: 420px 1fr;
-    align-items: flex-start;
+  .workspace__header {
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: stretch;
+    gap: 32px;
   }
+}
+
+.workspace__title {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  max-width: 720px;
+}
+
+.workspace__title h1 {
+  margin: 0;
+  font-size: 3rem;
+  letter-spacing: -0.03em;
+}
+
+.workspace__title p {
+  margin: 0;
+  color: rgba(226, 232, 240, 0.82);
+  line-height: 1.6;
+}
+
+.workspace__eyebrow {
+  margin: 0;
+  font-size: 0.75rem;
+  text-transform: uppercase;
+  letter-spacing: 0.22em;
+  color: rgba(148, 163, 184, 0.75);
+}
+
+.workspace__stats {
+  flex: 1;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(140px, 1fr));
+  gap: 16px;
+  align-content: start;
+}
+
+.stats-card {
+  background: rgba(15, 23, 42, 0.8);
+  border: 1px solid rgba(148, 163, 184, 0.28);
+  border-radius: 18px;
+  padding: 16px 18px;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  min-height: 92px;
+  box-shadow: 0 18px 40px rgba(15, 23, 42, 0.35);
+}
+
+.stats-card__label {
+  font-size: 0.8rem;
+  text-transform: uppercase;
+  letter-spacing: 0.12em;
+  color: rgba(148, 163, 184, 0.75);
+}
+
+.stats-card__value {
+  font-size: 1.4rem;
+  font-weight: 600;
+  color: #f8fafc;
+}
+
+.workspace__grid {
+  display: grid;
+  gap: 28px;
+}
+
+@media (min-width: 1080px) {
+  .workspace__grid {
+    grid-template-columns: minmax(0, 420px) minmax(0, 1fr);
+    align-items: start;
+  }
+}
+
+.workspace__column {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
 }
 
 .panel {
@@ -690,12 +748,6 @@ button.danger {
   border-radius: 999px;
   font-weight: 600;
   box-shadow: 0 18px 36px rgba(59, 130, 246, 0.35);
-}
-
-footer {
-  text-align: center;
-  color: rgba(148, 163, 184, 0.75);
-  font-size: 0.9rem;
 }
 
 .toast {
